Fixes to compilation with precompiled headers
